turbojpeg

Struct Compressor

pub struct Compressor { /* private fields */ }

Compresses raw pixel data into JPEG.

Implementations

impl Compressor

pub fn new() -> Result<Compressor>

Create a new compressor instance.

pub fn set_quality(&mut self, quality: i32) -> Result<()>

Set the quality of the compressed JPEG images.

The quality ranges from 1 (worst) to 100 (best).

Examples

let mut compressor = turbojpeg::Compressor::new()?;
compressor.set_quality(95)?;

An error is returned if quality is invalid:

let mut compressor = turbojpeg::Compressor::new()?;
assert!(compressor.set_quality(120).is_err());

pub fn set_lossless(&mut self, enabled: bool) -> Result<()>

Enable/disable lossless compression.

pub fn set_subsamp(&mut self, subsamp: Subsamp) -> Result<()>

Set the level of chrominance subsampling of the compressed JPEG images.

Chrominance subsampling can reduce the compressed image size without noticeable loss of quality (see Subsamp for more).

When compressing YUV images, the level of chrominance subsampling is set from the image, and the value set by this method is overriden.

pub fn set_optimize(&mut self, optimize: bool) -> Result<()>

Enable/disable optimized baseline entropy coding.

When enabled, optimal Huffman tables will be computed for the JPEG image. Optimized baseline entropy coding will improve compression slightly (generally 5% or less), but it will reduce compression performance considerably.

Example

let image = turbojpeg::Image::mandelbrot(500, 500, turbojpeg::PixelFormat::RGB);
let mut compressor = turbojpeg::Compressor::new()?;

compressor.set_optimize(false)?;
let unoptimized = compressor.compress_to_vec(image.as_deref())?;

compressor.set_optimize(true)?;
let optimized = compressor.compress_to_vec(image.as_deref())?;

assert!(optimized.len() < unoptimized.len());

pub fn compress( &mut self, image: Image<&[u8]>, output: &mut OutputBuf<'_>, ) -> Result<()>

Compresses the image into output buffer.

This is the main compression method, which gives you full control of the output buffer. If you don’t need this level of control, you can use one of the convenience wrappers below.

Example

// create an image (a Mandelbrot set visualization)
let image = turbojpeg::Image::mandelbrot(500, 500, turbojpeg::PixelFormat::RGB);

// initialize the compressor
let mut compressor = turbojpeg::Compressor::new()?;
compressor.set_quality(70);
compressor.set_subsamp(turbojpeg::Subsamp::Sub2x2);

// initialize the output buffer
let mut output_buf = turbojpeg::OutputBuf::new_owned();

// compress the image into JPEG
// (we use as_deref() to convert from &Image<Vec<u8>> to Image<&[u8]>)
compressor.compress(image.as_deref(), &mut output_buf)?;

// write the JPEG to disk
std::fs::write(std::env::temp_dir().join("mandelbrot.jpg"), &output_buf)?;

pub fn compress_to_owned(&mut self, image: Image<&[u8]>) -> Result<OwnedBuf>

Compresses the image into an owned buffer.

This method automatically allocates the memory for output and avoids needless copying.

pub fn compress_to_vec(&mut self, image: Image<&[u8]>) -> Result<Vec<u8>>

Compress the image into a new Vec<u8>.

This method copies the compressed data into a new Vec. If you would like to avoid the extra allocation and copying, consider using compress_to_owned() instead.

pub fn compress_to_slice( &mut self, image: Image<&[u8]>, output: &mut [u8], ) -> Result<usize>

Compress the image into the slice output.

Returns the size of the compressed JPEG data. If the compressed image does not fit into dest, this method returns an error. Use buf_len() to determine buffer size that is guaranteed to be large enough for the compressed image.

pub fn compress_yuv( &mut self, image: YuvImage<&[u8]>, output: &mut OutputBuf<'_>, ) -> Result<()>

Compresses the YuvImage into output buffer.

This is similar to compress(), but encodes a YUV image instead of RGB image. This method gives you full control of the output buffer. If you don’t need this level of control, you can use one of the convenience wrappers below.

Encoding YUV images is useful if you already have an image in YUV, for example, if you receive it from a camera.

Example

// parameters from rpi camera v3
const WIDTH: usize = 1536;
const HEIGHT: usize = 864;
const SIZE: usize = WIDTH * HEIGHT * 3 / 2;

// grab a raw yuv image from somewhere
let mut raw_yuv_pixels: Vec<u8> = vec![0; SIZE];

// initialize the compressor
let mut compressor = turbojpeg::Compressor::new()?;
compressor.set_quality(70);

// initialize the YuvImage
//  slice ref to avoid using as_deref
let yuv_image = turbojpeg::YuvImage{
    pixels: &raw_yuv_pixels[..],
    width: WIDTH,
    align: 4,
    height: HEIGHT,
    subsamp: turbojpeg::Subsamp::Sub2x2,
 };

// initialize the output buffer
let mut output_buf = turbojpeg::OutputBuf::new_owned();

// compress the image into JPEG
compressor.compress_yuv(yuv_image, &mut output_buf).unwrap();

// write the JPEG to disk
std::fs::write(std::env::temp_dir().join("tj_encoded.jpg"), &output_buf)?;

pub fn compress_yuv_to_owned( &mut self, image: YuvImage<&[u8]>, ) -> Result<OwnedBuf>

Compresses the YuvImage into an owned buffer.

This method automatically allocates the memory for output and avoids needless copying.

pub fn compress_yuv_to_vec(&mut self, image: YuvImage<&[u8]>) -> Result<Vec<u8>>

Compress the YuvImage into a new Vec<u8>.

This method copies the compressed data into a new Vec. If you would like to avoid the extra allocation and copying, consider using compress_yuv_to_owned() instead.

pub fn compress_yuv_to_slice( &mut self, image: YuvImage<&[u8]>, output: &mut [u8], ) -> Result<usize>

Compress the YuvImage into the slice output.

Returns the size of the compressed JPEG data. If the compressed image does not fit into dest, this method returns an error. Use compressed_buf_len() to determine buffer size that is guaranteed to be large enough for the compressed image.

pub fn compress_yuv_planes( &mut self, image: &YuvPlanesImage<&[u8]>, output: &mut OutputBuf<'_>, ) -> Result<()>

Compress separate YUV planes into JPEG.

This function compresses separate Y, U (Cb), and V (Cr) image planes into a JPEG image. This is similar to compress_yuv(), but takes separate YUV planes as input instead of interleaved YUV data.

Arguments

image - YUV image with separate planes

output - Output buffer for compressed JPEG data

Example

// prepare YUV plane data
let width = 640;
let height = 480;
let y_plane = vec![128u8; width * height];
let u_plane = vec![128u8; (width/2) * (height/2)];
let v_plane = vec![128u8; (width/2) * (height/2)];

let image = turbojpeg::YuvPlanesImage {
    y_plane: &y_plane[..],
    u_plane: &u_plane[..],
    v_plane: &v_plane[..],
    width,
    height,
    y_stride: width,
    u_stride: width/2,
    v_stride: width/2,
    subsamp: turbojpeg::Subsamp::Sub2x2,
};

// initialize the compressor
let mut compressor = turbojpeg::Compressor::new()?;
compressor.set_quality(90)?;

// initialize the output buffer
let mut output_buf = turbojpeg::OutputBuf::new_owned();

// compress YUV planes to JPEG
compressor.compress_yuv_planes(&image, &mut output_buf)?;

pub fn compress_yuv_planes_to_owned( &mut self, image: &YuvPlanesImage<&[u8]>, ) -> Result<OwnedBuf>

Compress separate YUV planes into an owned buffer.

This method automatically allocates the memory for output and avoids needless copying.

pub fn compress_yuv_planes_to_vec( &mut self, image: &YuvPlanesImage<&[u8]>, ) -> Result<Vec<u8>>

Compress separate YUV planes into a new Vec<u8>.

This method copies the compressed data into a new Vec. If you would like to avoid the extra allocation and copying, consider using compress_yuv_planes_to_owned() instead.

pub fn compress_yuv_planes_to_slice( &mut self, image: &YuvPlanesImage<&[u8]>, output: &mut [u8], ) -> Result<usize>

Compress separate YUV planes into the slice output.

Returns the size of the compressed JPEG data. If the compressed image does not fit into dest, this method returns an error. Use compressed_buf_len() to determine buffer size that is guaranteed to be large enough for the compressed image.

pub fn buf_len(&self, width: usize, height: usize) -> Result<usize>

Compute the maximum size of a compressed image.

This depends on image width and height, and also on the current setting of chrominance subsampling (see set_subsamp()).

You can also use compressed_buf_len() directly.