turbojpeg

Struct Decompressor

pub struct Decompressor { /* private fields */ }

Decompresses JPEG data into raw pixels.

Implementations

impl Decompressor

pub fn new() -> Result<Decompressor>

Create a new decompressor instance.

pub fn read_header(&mut self, jpeg_data: &[u8]) -> Result<DecompressHeader>

Read the JPEG header without decompressing the image.

Example

// read JPEG data from file
let jpeg_data = std::fs::read("examples/parrots.jpg")?;

// initialize a decompressor
let mut decompressor = turbojpeg::Decompressor::new()?;

// read the JPEG header
let header = decompressor.read_header(&jpeg_data)?;
assert_eq!((header.width, header.height), (384, 256));

pub fn set_scaling_factor( &mut self, scaling_factor: ScalingFactor, ) -> Result<()>

Set scaling factor for subsequent decompression operations.

Only the scaling factors returned by supported_scaling_factors() are supported, and a scaling factor can only be used when decompressing lossy JPEG images.

Example

// read JPEG data from file
let jpeg_data = std::fs::read("examples/parrots.jpg")?;

// initialize a decompressor with the scaling factor
let mut decompressor = turbojpeg::Decompressor::new()?;
let scaling = turbojpeg::ScalingFactor::ONE_HALF;
decompressor.set_scaling_factor(scaling);

// read the JPEG header and downscale the width and height
let scaled_header = decompressor.read_header(&jpeg_data)?.scaled(scaling);

// initialize the image (Image<Vec<u8>>)
let mut image = turbojpeg::Image {
    pixels: vec![0; 4 * scaled_header.width * scaled_header.height],
    width: scaled_header.width,
    pitch: 4 * scaled_header.width, // size of one image row in memory
    height: scaled_header.height,
    format: turbojpeg::PixelFormat::RGBA,
};

// decompress the JPEG into the image
// (we use as_deref_mut() to convert from &mut Image<Vec<u8>> into Image<&mut [u8]>)
decompressor.decompress(&jpeg_data, image.as_deref_mut())?;
assert_eq!(&image.pixels[0..5], &[125, 121, 92, 255, 127]);

pub fn scaling_factor(&self) -> ScalingFactor

Get the scaling factor set by set_scaling_factor().

pub fn decompress( &mut self, jpeg_data: &[u8], output: Image<&mut [u8]>, ) -> Result<()>

Decompress a JPEG image in jpeg_data into output.

The decompressed image is stored in the pixel data of the given output image, which must be fully initialized by the caller. Use read_header() to determine the image size before calling this method.

Example

// read JPEG data from file
let jpeg_data = std::fs::read("examples/parrots.jpg")?;

// initialize a decompressor
let mut decompressor = turbojpeg::Decompressor::new()?;

// read the JPEG header
let header = decompressor.read_header(&jpeg_data)?;

// initialize the image (Image<Vec<u8>>)
let mut image = turbojpeg::Image {
    pixels: vec![0; 4 * header.width * header.height],
    width: header.width,
    pitch: 4 * header.width, // size of one image row in memory
    height: header.height,
    format: turbojpeg::PixelFormat::RGBA,
};

// decompress the JPEG into the image
// (we use as_deref_mut() to convert from &mut Image<Vec<u8>> into Image<&mut [u8]>)
decompressor.decompress(&jpeg_data, image.as_deref_mut())?;
assert_eq!(&image.pixels[0..4], &[122, 118, 89, 255]);

pub fn decompress_to_yuv( &mut self, jpeg_data: &[u8], output: YuvImage<&mut [u8]>, ) -> Result<()>

Decompress a JPEG image in jpeg_data into output as YUV without changing color space.

The decompressed image is stored in the pixel data of the given output image, which must be fully initialized by the caller. Use read_header() to determine the image size before calling this method.

Example

// read JPEG data from file
let jpeg_data = std::fs::read("examples/parrots.jpg")?;

// initialize a decompressor
let mut decompressor = turbojpeg::Decompressor::new()?;

// read the JPEG header
let header = decompressor.read_header(&jpeg_data)?;
// calculate YUV pixels length
let align = 4;
let yuv_pixels_len = turbojpeg::yuv_pixels_len(header.width, align, header.height, header.subsamp);

// initialize the image (YuvImage<Vec<u8>>)
let mut image = turbojpeg::YuvImage {
    pixels: vec![0; yuv_pixels_len.unwrap()],
    width: header.width,
    align,
    height: header.height,
    subsamp: header.subsamp,
};

// decompress the JPEG into the image
// (we use as_deref_mut() to convert from &mut YuvImage<Vec<u8>> into YuvImage<&mut [u8]>)
decompressor.decompress_to_yuv(&jpeg_data, image.as_deref_mut())?;
assert_eq!(&image.pixels[0..4], &[116, 117, 118, 119]);

pub fn supported_scaling_factors() -> Vec<ScalingFactor>

Get the list of scaling factors supported for decompression.

At the time of this writing, TurboJPEG supports all multiples of 1/8 between 1/8 and 2 as scaling factors, but it’s unclear whether this is guaranteed to continue to be the case in all future versions.

Example

let factors = turbojpeg::Decompressor::supported_scaling_factors();
for num in 1..16 {
    let multiple_of_8 = turbojpeg::ScalingFactor::new(num, 8);
    assert!(factors.iter().find(|&f| *f == multiple_of_8).is_some());
}