turbojpeg

Struct Transformer

pub struct Transformer { /* private fields */ }

Transforms JPEG images without recompression.

TurboJPEG applies the transformation on the DCT coefficients, without performing complete decompression. This is faster and also means that the transforms are lossless.

Implementations

impl Transformer

pub fn new() -> Result<Transformer>

Create a new transformer instance.

pub fn transform( &mut self, transform: &Transform, jpeg_data: &[u8], output: &mut OutputBuf<'_>, ) -> Result<()>

Apply a transformation to the compressed JPEG.

This is the main transformation method, which gives you full control of the output buffer. If you don’t need this level of control, you can use one of the convenience wrappers below.

Example

// read JPEG data from file
let jpeg_data = std::fs::read("examples/parrots.jpg")?;

// initialize the transformer
let mut transformer = turbojpeg::Transformer::new()?;

// define the transformation: flip vertically, trim partial MCU blocks on the bottom edge 
let mut transform = turbojpeg::Transform::op(turbojpeg::TransformOp::Vflip);
transform.trim = true;

// initialize the output buffer
let mut flipped_data = turbojpeg::OutputBuf::new_owned();

// apply the transformation
transformer.transform(&transform, &jpeg_data, &mut flipped_data)?;

// write the flipped JPEG back to disk
std::fs::write(std::env::temp_dir().join("flipped_parrots.jpg"), &flipped_data)?;

pub fn transform_to_owned( &mut self, transform: &Transform, jpeg_data: &[u8], ) -> Result<OwnedBuf>

Transforms the image into an owned buffer.

This method automatically allocates the memory and avoids needless copying.

pub fn transform_to_vec( &mut self, transform: &Transform, jpeg_data: &[u8], ) -> Result<Vec<u8>>

Transform the image into a new Vec<u8>.

This method copies the transformed data into a new Vec. If you would like to avoid the extra allocation and copying, consider using transform_to_owned() instead.

pub fn transform_to_slice( &mut self, transform: &Transform, jpeg_data: &[u8], output: &mut [u8], ) -> Result<usize>

Transform the image into the slice output.

Returns the size of the transformed JPEG data. If the transformed image does not fit into dest, this method returns an error.

You can use compressed_buf_len() to determine buffer size that should be enough for the image, but there are some rare cases (such as transforming images with a large amount of embedded EXIF or ICC profile data) in which the output image will be larger than the size returned by compressed_buf_len().